requirements.txt
aiogram==3.20.0.post0
python-dotenv==1.1.0
